Added samples
