+++ sdk/mixedreality/mixed-reality-authentication/samples/javascript/package.json
@@ -1,15 +1,15 @@
 {
-  "name": "azure-template-samples-js",
+  "name": "azure-mixed-reality-authentication-samples-js",
   "private": true,
   "version": "0.1.0",
-  "description": "Azure template client library samples for JavaScript",
+  "description": "Azure Mixed Reality Authentication client library samples for JavaScript",
   "engine": {
     "node": ">=8.0.0"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
-    "directory": "sdk/template/template"
+    "directory": "sdk/mixedreality/mixed-reality-authentication"
   },
   "keywords": [
     "Azure",
@@ -21,10 +21,10 @@
   "bugs": {
     "url": "https://github.com/Azure/azure-sdk-for-js/issues"
   },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/template/template",
+  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/mixedreality/mixed-reality-authentication",
   "sideEffects": false,
   "dependencies": {
-    "@azure/template": "latest",
+    "@azure/mixed-reality-authentication": "latest",
     "dotenv": "^8.2.0"
   }
 }
