+++ sdk/mixedreality/mixed-reality-authentication/samples/typescript/README.md
@@ -6,17 +6,17 @@ languages:
   - typescript
 products:
   - azure
-  - azure-template
-urlFragment: template-typescript
+  - mixed-reality-authentication
+urlFragment: mixed-reality-authentication-typescript
 --- -->
 
-# Azure Template client library samples for TypeScript
+# Azure Mixed Reality Authentication client library samples for TypeScript
 
 These sample programs show how to use the TypeScript client libraries for Azure Template in some common scenarios.
 
-| **File Name**                       | **Description** |
-| ----------------------------------- | --------------- |
-| [sampleTemplate.ts][sampletemplate] | sample template |
+| **File Name**           | **Description** |
+| ----------------------- | --------------- |
+| [getToken.ts][getToken] | Get a token.    |
 
 ## Prerequisites
 
@@ -53,21 +53,21 @@ npm run build
 4. Run whichever samples you like (note that some samples may require additional setup, see the table above):
 
 ```bash
-node dist/sampleTemplate.js
+node dist/getToken.js
 ```
 
 Alternatively, run a single sample with the correct environment variables set (step 3 is not required if you do this), for example (cross-platform):
 
 ```bash
-npx cross-env ENDPOINT="<endpoint>" API_KEY="<api key>" node dist/sampleTemplate.js
+npx cross-env MIXEDREALITY_ACCOUNT_DOMAIN="<account domain>" MIXEDREALITY_ACCOUNT_ID="<account identifier>" MIXEDREALITY_ACCOUNT_KEY="<account key>" node dist/getToken.js
 ```
 
 ## Next Steps
 
 Take a look at our [API Documentation][apiref] for more information about the APIs that are available in the clients.
 
-[sampletemplate]: https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/template/template/samples/typescript/src/sampleTemplate.ts
+[getToken]: https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/mixedreality/mixed-reality-authentication/samples/typescript/src/getToken.ts
 [apiref]: https://docs.microsoft.com/javascript/api
 [freesub]: https://azure.microsoft.com/free/
-[package]: https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/template/template/README.md
+[package]: https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/mixedreality/mixed-reality-authentication/README.md
 [typescript]: https://www.typescriptlang.org/docs/home.html
