+++ sdk/mixedreality/mixed-reality-authentication/samples/typescript/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "azure-template-samples-ts",
+  "name": "azure-mixed-reality-authentication-samples-ts",
   "private": true,
   "version": "0.1.0",
-  "description": "Azure template client library samples for TypeScript",
+  "description": "Azure Mixed Reality Authentication client library samples for TypeScript",
   "engine": {
     "node": ">=8.0.0"
   },
@@ -13,7 +13,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
-    "directory": "sdk/template/template"
+    "directory": "sdk/mixedreality/mixed-reality-authentication"
   },
   "keywords": [
     "Azure",
@@ -28,7 +28,7 @@
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/template/template",
   "sideEffects": false,
   "dependencies": {
-    "@azure/template": "latest",
+    "@azure/mixed-reality-authentication": "latest",
     "dotenv": "^8.2.0"
   },
   "devDependencies": {
