+++ sdk/mixedreality/mixed-reality-authentication/samples/typescript/sample.env
@@ -1,7 +1,8 @@
 # Used in most samples. Retrieve these values from a Cognitive Services instance
 # in the Azure Portal.
-ENDPOINT="Your Endpoint URL"
-API_KEY="<Your API key>"
+MIXEDREALITY_ACCOUNT_DOMAIN=<MR account domain>
+MIXEDREALITY_ACCOUNT_ID=<MR account identifier>
+MIXEDREALITY_ACCOUNT_KEY=<MR account key>
 
 # Used to authenticate using Azure AD as a service principal for role-based authentication
 # in the tokenAuth sample.
